+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -22,6 +22,7 @@
 ## Version 7.7.1
 * Updated `GetAzureRmRecoveryServicesBackupSchedulePolicyObject` to default the `PolicySubType` to `Enhanced` for AzureVM.
 * Updated Azure Backup SDK version to support protecting TVM with standard policy.
+* Added fix for resume protection with AzureFileshare.
 
 ## Version 7.7.0
 * Fix for reprotect cmdlet in Azure Site Recovery for Azure to Azure provider.